commit 1.7 Updated Generate Tasks

+++ src/main/java/org/example/kihelp/task/mapper/impl/TaskToTaskResponseMapperImpl.java
@@ -17,7 +17,7 @@ public class TaskToTaskResponseMapperImpl implements TaskToTaskResponseMapper {
                 task.getId(),
                 subTitle,
                 task.getTitle(),
-                task.getPrice(),
+                (double)task.getPrice(),
                 task.getDiscount(),
                 task.getType(),
                 task.getTeacher().getId(),
